refactor: remove components/forms/destination
