+++ src/components/destination/index.ts
@@ -1,5 +1,7 @@
 import ConfigureDestinationForm from "./ConfigureDestinationForm";
 import type { ConfigureDestinationFormProps } from "./ConfigureDestinationForm";
+import CreateDestinationForm from "./CreateDestinationForm";
+import type { CreateDestinationFormProps } from "./CreateDestinationForm";
 
-export { ConfigureDestinationForm };
-export type { ConfigureDestinationFormProps };
+export { ConfigureDestinationForm, CreateDestinationForm };
+export type { ConfigureDestinationFormProps, CreateDestinationFormProps };
