--- src/components/forms/connector/destination/ConfigureDestinationForm/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ConfigureDestinationForm";
-export * from "./ConfigureDestinationForm";
