--- src/components/forms/connector/destination/CreateDestinationForm/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./CreateDestinationForm";
-export * from "./CreateDestinationForm";
