--- src/components/forms/connector/destination/index.ts
@@ -1,14 +1,0 @@
-import CreateDestinationForm from "./CreateDestinationForm";
-import type { CreateDestinationFormProps } from "./CreateDestinationForm";
-import ConfigureDestinationForm from "./ConfigureDestinationForm";
-import type {
-  ConfigureDestinationFormProps,
-  ConfigureDestinationFormValue,
-} from "./ConfigureDestinationForm";
-
-export { CreateDestinationForm, ConfigureDestinationForm };
-export type {
-  CreateDestinationFormProps,
-  ConfigureDestinationFormProps,
-  ConfigureDestinationFormValue,
-};
