+++ src/components/forms/connector/index.ts
@@ -1,2 +1,1 @@
 export * from "./source";
-export * from "./destination";
