+++ src/components/forms/pipeline/CreatePipelineForm/SetupDestinationStep/SetupDestinationStep.tsx
@@ -17,7 +17,7 @@ import { useCreateDestination, useDestinations } from "@/services/connector";
 import { CreateDestinationPayload } from "@/lib/instill";
 import { useAmplitudeCtx } from "@/contexts/AmplitudeContext";
 import { sendAmplitudeData } from "@/lib/amplitude";
-import { CreateDestinationForm } from "@/components/forms/connector";
+import { CreateDestinationForm } from "@/components/destination";
 
 export type SetupDestinationStepProps = StepNumberState;
